Comment-update

+++ FrontEnd/assets/javascript/modal.js
@@ -9,37 +9,45 @@ const btnAddPhoto = document.querySelector(".btn-add-photo");
 const btnValid = document.querySelector(".btn-valid");
 const btnDeleteAll = document.querySelector(".btn-delete-all");
 
+// Function for open the modal
 modalLink.onclick = function () {
   modal.style.display = null;
 };
 
+// Function for close the modal
 btnClose.onclick = function () {
   modal.style.display = "none";
 };
 
+// Function for open the Add photo modal
 btnAddPhoto.onclick = function () {
   createAddPhotoModal();
 };
 
+// Function for return to the previous modal
 btnReturn.onclick = function () {
   container.innerHTML = "";
   createGalerieWork();
 };
 
+// Function for add work
 btnValid.addEventListener("click", addWork);
 
+// Function to close the modal on click
 window.onclick = function (e) {
   if (e.target == modal) {
     modal.style.display = "none";
   }
 };
 
+// Function to close the modal with ESC
 window.addEventListener("keydown", function (e) {
   if (e.key === "Escape" || e.key === "Esc") {
     modal.style.display = "none";
   }
 });
 
+// Function to create the gallery modal
 function createGalerieModal() {
   container.style.display = "grid";
   titleModal.textContent = "Galerie photo";
@@ -51,6 +59,7 @@ function createGalerieModal() {
   btnAddPhoto.style.display = "block";
 }
 
+// Function to display the works on the modal gallery
 function createGalerieWork() {
   createGalerieModal();
   works.forEach((article) => {
@@ -63,6 +72,7 @@ function createGalerieWork() {
     editElement.innerText = "éditer";
     let flexButton = document.createElement("div");
     flexButton.classList.add("flex-button");
+    // Create of the button to delete work
     let btnDelete = document.createElement("button");
     btnDelete.classList.add("btn-delete");
     btnDelete.innerHTML = '<i class="fa-solid fa-trash-can"></i>';
@@ -73,15 +83,19 @@ function createGalerieWork() {
     articleElement.addEventListener("mouseenter", () => {
       btnExtend.style.opacity = "1";
     });
-
     articleElement.addEventListener("mouseleave", () => {
       btnExtend.style.opacity = "0";
     });
 
+    // Confirmation before delete works
     btnDelete.addEventListener("click", () => {
-      deleteWork(article.id);
+      if (confirm("voulez-vous supprimmer?")) {
+        deleteWork(article.id);
+      } else {
+        return;
+      }
     });
-
+    // Display all the elements
     articleElement.append(flexButton);
     flexButton.append(btnExtend);
     flexButton.append(btnDelete);
@@ -91,6 +105,7 @@ function createGalerieWork() {
   });
 }
 
+// Fonction for delete work
 function deleteWork(id) {
   fetch(`http://localhost:5678/api/works/${id}`, {
     method: "delete",
@@ -98,11 +113,12 @@ function deleteWork(id) {
       Authorization: `bearer ${localStorage.token}`,
     },
   })
+    // API response management
     .then((response) => {
       if (response.ok) {
-        alert("suprimer");
+        alert("Projet suprimé");
       } else if (response.status === 401) {
-        alert("Vous n'êtes pas authorisé à modifier le contenu");
+        alert("Vous n'êtes pas autorisé à modifier le contenu");
       } else if (response.status === 500) {
         alert("Problème avec le serveur");
       } else {
@@ -114,6 +130,7 @@ function deleteWork(id) {
     });
 }
 
+// Function to creat the add project modal
 function createAddPhotoModal() {
   createCategory();
   container.style.display = "flex";
@@ -122,8 +139,10 @@ function createAddPhotoModal() {
   btnReturn.style.display = "block";
   btnDeleteAll.style.display = "none";
   btnValid.style.display = "block";
+  btnValid.style.backgroundColor = "grey";
   btnAddPhoto.style.display = "none";
   navModal.style.justifyContent = "space-between";
+  // Create the form for add the new project
   container.innerHTML = `
     <form action="#" method="post">
     <div class="photo-preview">
@@ -144,6 +163,7 @@ function createAddPhotoModal() {
     </form>`;
 }
 
+// Function to display a preview of the photo uploaded
 function displayPhoto() {
   const file = document.querySelector("#upload-photo").files[0];
   const photoWrapper = document.querySelector(".upload-container");
@@ -161,6 +181,7 @@ function displayPhoto() {
   }
 }
 
+// Function to create the category option of the form
 function createCategory() {
   fetch("http://localhost:5678/api/categories")
     .then((res) => res.json())
@@ -176,41 +197,45 @@ function createCategory() {
     });
 }
 
+// Function to add the new project from the form
 function addWork() {
   const uploaded = document.querySelector("#upload-photo").files[0];
   const title = document.querySelector("#title").value;
   const category = document.querySelector("#category").value;
+  // Create a form data for the API
   const formData = new FormData();
   formData.append("image", uploaded);
   formData.append("title", title);
   formData.append("category", category);
-
-  console.log(uploaded);
-  console.log(category);
-  console.log(title);
-
-  fetch("http://localhost:5678/api/works", {
-    method: "post",
-    body: formData,
-    headers: {
-      // 'Content-Type': 'multipart/form-data',
-      Accept: "application/json",
-      Authorization: `bearer ${localStorage.token}`,
-    },
-  })
-    .then((response) => {
-      if (response.ok) {
-        alert("photo ajouter");
-      } else if (response.status === 400) {
-        alert(`formulaire incomplet`);
-      } else if (response.status === 401) {
-        alert(`utilisateur non authoriser`);
-      } else if (response.status === 500) {
-        alert(`error inattendue`);
-      } else {
-        alert(`erreur: ${response.status}`);
-      }
+  // Check if all the input are filled
+  if (!uploaded || !title || !category) {
+    alert("Veuillez remplir tous les champs");
+    return;
+  } else {
+    // Send the form data with POST with the token
+    fetch("http://localhost:5678/api/works", {
+      method: "post",
+      body: formData,
+      headers: {
+        Accept: "application/json",
+        Authorization: `bearer ${localStorage.token}`,
+      },
     })
-    .then((data) => console.log(data))
-    .catch((error) => console.log(error));
+      // API response management
+      .then((response) => {
+        if (response.ok) {
+          alert("Projet ajouté");
+        } else if (response.status === 400) {
+          alert("Formulaire incomplet");
+        } else if (response.status === 401) {
+          alert("utilisateur non authoriser");
+        } else if (response.status === 500) {
+          alert("erreur inattendue");
+        } else {
+          alert(`erreur: ${response.status}`);
+        }
+      })
+      .then((data) => console.log(data))
+      .catch((error) => console.log(error));
+  }
 }
